+++ src/main/java/com/kk/newSubway/repository/StationRepository.java
@@ -3,5 +3,7 @@ package com.kk.newSubway.repository;
 import com.kk.newSubway.model.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StationRepository extends JpaRepository<Station, Long> {
+public interface StationRepository extends JpaRepository<Station, Integer> {
+
+
 }
