+++ src/main/java/com/kk/newSubway/repository/ZoneRepository.java
@@ -3,5 +3,5 @@ package com.kk.newSubway.repository;
 import com.kk.newSubway.model.Zone;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ZoneRepository extends JpaRepository<Zone, Long> {
+public interface ZoneRepository extends JpaRepository<Zone, Integer> {
 }
